move interop to a new module
